requirements.txt
djoser
Django==3.2.7
djangorestframework==3.13.1
Pillow